server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pandas
scikit-learn
httpx
pytest
pytest-asyncio
